Move FileFinderISISInstrumentDataCacheTest to it's own file so it actually runs
